Permit to bind object-server on replication_port
Add a new configuration option 'servers_per_port_type' to choose on which
port the object-server must bind when servers_per_port is enabled. It can
be 'port' (the default, which does not change the current behavior), or
'replication_port', or both separated by a comma.

The following configuration will create two workers per device if port
and replication_port are different in the ring. One worker for clients'
traffic and one worker for replication traffic:
    servers_per_port = 1
    servers_per_port_type = port, replication_port

If using different ports in the ring for proxy and replication traffic,
it allows to dedicate processes to replication, which can be
IO-intensive and sometime hangs the process for many seconds.

Closes-Bug: #1669579

Change-Id: I991806227f18f115646b0549a77326617aa2c8fe

Change-Id: I3f4e986767032c7c9e652ccea6203f90aef74924