crypto: do not fail on missing key when in SSE-C mode
If there is an If-Match or If-None-Match header in a GET or HEAD
request, the encryption middleware will try to load encryption keys
before doing the GET or HEAD subrequest. When the SSE-C mode is enabled,
the keys may not be available (because they are provided by the
requester) but we should not consider that an internal error.

Change-Id: Ife01f0fadb21380fee9bf27a280e9e885be4cccc
Topic: OpenIO backend