s3api: Implement a bucket logging middleware
S3 commands, visible in the `log_info` fields, have been replaced
by S3 operations (format used by AWS in their logs).
Here are some examples of this replacement:
  list-objects         -> REST.GET.BUCKET
  list-object-versions -> REST.GET.BUCKETVERSIONS
  delete-bucket-cors   -> REST.DELETE.CORS
  head-object          -> REST.HEAD.BUCKET
  put-object-tagging   -> REST.PUT.TAGGING

Now, when one of the controllers is called, the S3 operation is
immediately filled in order to always have the information,
even in the event of an error during the first checks (decorators).

Change-Id: Ib4f60b51bf90dd91663e2ef4723727068727954b