s3api: Ignore Content-MD5 for copy-object request
Some clients(MinioJava) may send a 'Content-MD5' header for copy-object
request. This behavior leads to reject copy request for non matching ETags since
'Content-MD5' header is converted to 'ETag' header.

The 'Content-MD5' is now droped before conversion to 'ETag'if the header
'X-Amz-Copy-Source' is present. This behavior matchs the Amazon one.

Change-Id: I7f3181592e42aa7e5ea523c938fdd4c38407cb48