s3api: Immediately deny access when user policies do not explicitly allow
As with AWS, user policies give the user the right to use an operation.
But to actually perform the operation, the user must have rights
to the resource through ACLs (or bucket policies, when they will be
available).

Additionally, when user policies are enabled, permission to create
a bucket is entirely managed by these user policies. User roles
in the authentication system are therefore ignored.

Change-Id: If5129618ac6719db54818ed0e60690a045d44eb4