intelligent-tiering: Reduce the likelihood of launching multiple archives
When an archive is requested, checking the bucket can take a long time,
especially checking for uncompleted MPUs.
During these checks, the client can (re)request to archive the same bucket
(because the response is too long).

If the status check is done too soon, we send multiple messages in RabbitMQ.
By checking the status as late as possible, we are more likely
to deny next requests.

This issue also exists for restoring and deleting the bucket
(to a lesser extent).

Change-Id: I3f9345e989dd8fbc9670ccc9816c6ebd56829407