s3api: Create a true new version when copying an object
Since enabling the heartbeat when copying on the server side
no longer worked due to self-copying for MPUs.

Additionally, if versioning is enabled, it is important not
to have 2 versions of objects pointing to the same parts.
Otherwise deleting one of the versions will delete all parts.

Also, do not copy MPU related properties into the new object.
The new object is no longer an MPU, but a simple object.

Change-Id: I3727802beca812ec3821780e76dbe424c41a4a85
Co-Authored-By: Romain de Joux <romain.de-joux@ovhcloud.com>