Optimize how flexible s3token cache duration
Currently s3token `secret_cache_duration_min` is used to modulate cache
duration between this value and `secret_cache_duration`. That allow to
spread Keystone requests after a cache reset. This solution doesn't
protect Keystone about burst of request on cache invalidation in case of
high load from a unique client (more than 10K req/s for ex.).

This commit modify how we invalidate cache. The idea is to always set
memcached ttl to `secret_cache_duration` and to randomly ignore it on a
client request between `secret_cache_duration_min` and
`secret_cache_duration`.

To do that, we need to store the timestamp of the end of the cache
validity in cached value (memcached don't provide this value), to allow
us to compute the current_ttl of the value. With a ratio between
current_ttl and delta_ttl (from config) we can randomly ignore the cache
and overwrite the current one.