mpu: do not require a key during SSE-C CompleteMultiPartUpload
awscli sends a key during CompleteMultiPartUpload, rclone does not. When
encyption is enabled on the bucket, but no key is provided in the
request, we will try to decode the ETag with the bucket's key, and fail.

With this change we allow CompleteMultiPartUpload without key. The
previous behavior can be re-established by setting
complete_without_key=False in encryption middleware configuration.

Change-Id: I5040c5805d6ecbd16b47afad4e3d95422e523626