s3api: Map INTELLIGENT_TIERING with the STANDARD offer
The INTELLIGENT_TIERING storage class is a very specific class that allows
storage to be scaled based on the last access to the object, starting with
frequently used storage (equivalent to STANDARD).
If an S3 client uses this storage class, even on a non-standard endpoint,
they will expect to be able to access it frequently (at least for the first
30 days).
However, until this storage class is managed, the client will not benefit
from the automatic cost reduction.

Change-Id: I8340d81b4025a63aacae44845c76394d7d5ee218